bot/BOOM/requirements.txt
pandas
numpy
ccxt
ta-lib
python-dotenv
prometheus_client
requests
